feat:통계페이지, 카드추천페이지 쿼리문, 레파지토리 , 서비스, 컨트롤러, 뷰 수정

+++ src/main/java/com/oopsw/clario/service/card/CardService.java
@@ -3,9 +3,7 @@ package com.oopsw.clario.service.card;
 import com.oopsw.clario.dto.card.AllCardsDTO;
 import com.oopsw.clario.dto.statistics.CategoryStatisticsDTO;
 import com.oopsw.clario.dto.statistics.MonthlyExpenseComparisonDTO;
-import com.oopsw.clario.dto.statistics.Top3CategoriesDTO;
 import com.oopsw.clario.repository.card.CardRepository;
-import com.oopsw.clario.repository.statistics.StatisticsRepository;
 import com.oopsw.clario.service.statistics.StatisticsService;
 import com.oopsw.clario.util.CategoryMapper;
 import lombok.RequiredArgsConstructor;
@@ -13,7 +11,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 
 @RequiredArgsConstructor
@@ -42,15 +39,15 @@ public class CardService {
         return cardRepository.getCardsByCategoryNamesAndType(subCategories, cardType);
     }
 
-    public CategoryStatisticsDTO getCategoryStatistics(Long memberId, Long year, Long month) {
+    public CategoryStatisticsDTO getCategoryStatistics(Integer memberId, Long year, Long month) {
         return statisticsService.getCategoryStatistics(memberId, year, month);
     }
 
-    public Long getMonthlyExpenseSum(Long memberId, Long year, Long month) {
+    public Long getMonthlyExpenseSum(Integer memberId, Long year, Long month) {
         return statisticsService.getMonthlyExpenseSum(memberId, year, month);
     }
 
-    public MonthlyExpenseComparisonDTO getMonthlyExpenseComparison(Long memberId, Long year, Long month) {
+    public MonthlyExpenseComparisonDTO getMonthlyExpenseComparison(Integer memberId, Long year, Long month) {
         return statisticsService.getMonthlyExpenseComparison(memberId, year, month);
     }
