+++ src/main/java/com/oopsw/clario/service/statistics/StatisticsService.java
@@ -3,6 +3,7 @@ package com.oopsw.clario.service.statistics;
 import com.oopsw.clario.dto.statistics.*;
 import com.oopsw.clario.repository.statistics.StatisticsRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Service;
 
 import java.util.*;
@@ -22,7 +23,7 @@ public class StatisticsService {
 //        return result;
 //    }
 
-    public List<MonthlyExpenseTotalDTO> getMonthlyExpenseTotal(Long memberId, Long year, Long month) {
+    public List<MonthlyExpenseTotalDTO> getMonthlyExpenseTotal(Integer memberId, Long year, Long month) {
         List<MonthlyExpenseTotalDTO> result = statisticsRepository.getMonthlyExpenseTotal(memberId, year, month);
         if (result == null || result.isEmpty()) {
             System.out.println("⚠ memberId=" + memberId + ", year=" + year + ", month=" + month + " 에 대한 소비 내역이 없습니다.");
@@ -32,7 +33,7 @@ public class StatisticsService {
     }
 
 
-    public List<Top3CategoriesDTO> getTop3Categories(Long memberId, Long year, Long month) {
+    public List<Top3CategoriesDTO> getTop3Categories(Integer memberId, Long year, Long month) {
         List<Top3CategoriesDTO> result = statisticsRepository.getTop3Categories(memberId, year, month);
 
         if (result == null || result.isEmpty()) {
@@ -44,7 +45,7 @@ public class StatisticsService {
     }
 
 
-    public List<MonthlyCardTradeTotalDTO> getMonthlyCardTradeTotal(Long memberId, Long year) {
+    public List<MonthlyCardTradeTotalDTO> getMonthlyCardTradeTotal(Integer memberId, Long year) {
         List<MonthlyCardTradeTotalDTO> result = statisticsRepository.getMonthlyCardTradeTotal(memberId, year);
         if (result == null || result.isEmpty()) {
             System.out.println("⚠ memberId=" + memberId + " 에 대한 소비 내역이 없습니다.");
@@ -53,7 +54,7 @@ public class StatisticsService {
         return result;
     }
 
-    public List<YearlyExpenseDTO> getYearlyTotalExpense(Long memberId) {
+    public List<YearlyExpenseDTO> getYearlyTotalExpense(Integer memberId) {
         List<YearlyExpenseDTO> result = statisticsRepository.getYearlyTotalExpense(memberId);
         if (result == null || result.isEmpty()) {
             System.out.println("⚠ memberId=" + memberId + " 에 대한 연간 소비 내역이 없습니다.");
@@ -62,7 +63,7 @@ public class StatisticsService {
         return result;
     }
 
-    public List<YearlyIncomeDTO> getYearlyTotalIncome(Long memberId) {
+    public List<YearlyIncomeDTO> getYearlyTotalIncome(Integer memberId) {
         List<YearlyIncomeDTO> cardList = statisticsRepository.getYearlyIncomeFromCard(memberId);
         List<YearlyIncomeDTO> accountList = statisticsRepository.getYearlyIncomeFromAccount(memberId);
 
@@ -81,19 +82,19 @@ public class StatisticsService {
                 .collect(Collectors.toList());
     }
 
-    public List<MonthlyIncomeAverageDTO> getMonthlyAverageIncome(Long memberId) {
+    public List<MonthlyIncomeAverageDTO> getMonthlyAverageIncome(Integer memberId) {
         return statisticsRepository.getMonthlyAverageIncome(memberId);
     }
 
-    public List<MonthlyExpenseAverageDTO> getMonthlyExpenseAverage(Long memberId) {
+    public List<MonthlyExpenseAverageDTO> getMonthlyExpenseAverage(Integer memberId) {
         return statisticsRepository.getMonthlyExpenseAverage(memberId);
     }
 
-    public MonthlyIncomeDTO getMonthlyIncome(Long memberId, Long year, Long month) {
+    public MonthlyIncomeDTO getMonthlyIncome(Integer memberId, Long year, Long month) {
         return statisticsRepository.getMonthlyIncome(memberId, year, month);
     }
 
-    public CategoryStatisticsDTO getCategoryStatistics(Long memberId, Long year, Long month) {
+    public CategoryStatisticsDTO getCategoryStatistics(Integer memberId, Long year, Long month) {
         List<TopCategoryByCountDTO> countBased = statisticsRepository.getTopCategoriesByCount(memberId, year, month);
         List<TopCategoryByAmountDTO> amountBased = statisticsRepository.getTopCategoriesByAmount(memberId, year, month);
 
@@ -103,11 +104,11 @@ public class StatisticsService {
         return dto;
     }
 
-    public Long getMonthlyExpenseSum(Long memberId, Long year, Long month) {
+    public Long getMonthlyExpenseSum(Integer memberId, Long year, Long month) {
         return statisticsRepository.getMonthlyExpenseSum(memberId, year, month);
     }
 
-    public Double getMonthlyExpenseGrowthRate(Long memberId, Long year, Long month) {
+    public Double getMonthlyExpenseGrowthRate(Integer memberId, Long year, Long month) {
         // 현재 월 출금 합계
         Long current = statisticsRepository.getMonthlyExpenseSum(memberId, year, month);
         if (current == null) current = 0L;
@@ -132,7 +133,7 @@ public class StatisticsService {
         return Math.round(growthRate * 100.0) / 100.0;
     }
 
-    public MonthlyExpenseComparisonDTO getMonthlyExpenseComparison(Long memberId, Long year, Long month) {
+    public MonthlyExpenseComparisonDTO getMonthlyExpenseComparison(Integer memberId, Long year, Long month) {
         Long current = statisticsRepository.getMonthlyExpenseSum(memberId, year, month);
         if (current == null) current = 0L;
 
@@ -159,7 +160,7 @@ public class StatisticsService {
         return new MonthlyExpenseComparisonDTO(current, previous, rate);
     }
 
-    public IncomeVsExpenseDTO getIncomeVsExpense(Long memberId, Long year, Long month) {
+    public IncomeVsExpenseDTO getIncomeVsExpense(Integer memberId, Long year, Long month) {
         Long income = statisticsRepository.getMonthlyIncomeSum(memberId, year, month);
         Long expense = statisticsRepository.getMonthlyExpenseSum(memberId, year, month);
